Create README.md
The user will click on one of the grid boxes out of the nine available options initially. There will be a message printed below the pygame window which will state if it is X's turn or O's turn. This will be followed alternately till any one side wins. A message below the pygame window will be printed stating the winner. A draw message will be printed if neither side wins.